Seperated parse functions. Added some new ones.
