+++ clientSide/src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
-import { initializeParse } from "./parse-functions/parseFunctions";
+import { initializeParse } from "./parse-functions/parseInitializerFunction";
 
 initializeParse();
